requirements.txt
mss
pytesseract
deep-translator
pystray
Pillow
keyboard
screeninfo
